ACUMORG-53: Remove breaking characters from script
Checkmark and cross mark are breaking the script when executed on
machines with encoding different from the author's machine. As they are
not essential for the process, it was simpler to remove them than to
find a substitute or add instructions to ensure they are properly
encoded in the script.